Optionally transmit the audio data of a detected wake word.
The intention is to capture false positives, e.g. the user could respond
with "this was not the wake word" or simply "get away" to indicate a
wake-word failure. Some mechanism could the store away the audio data to
use it later as "not wake-word" in order to train a wake-word software.